+++ mate-panel/panel-ditem-editor.c
@@ -1309,7 +1309,7 @@ response_cb (GtkDialog *dialog,
 static void
 panel_ditem_editor_revert (PanelDItemEditor *dialog)
 {
-	int       i;
+	gsize     i;
 	char     *string;
 	gboolean  boolean;
 	GKeyFile *key_file;
@@ -1385,7 +1385,7 @@ panel_ditem_editor_revert (PanelDItemEditor *dialog)
 static void
 panel_ditem_editor_set_revert (PanelDItemEditor *dialog)
 {
-	int       i;
+	gsize     i;
 	char     *string;
 	gboolean  boolean;
 	GKeyFile *key_file;
